requirements2.txt
Flask
Flask-CORS
python-dotenv
Werkzeug
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-chroma
chromadb
pydantic
huggingface-hub
transformers
torch
sentence-transformers

flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
langchain==0.0.267
langchain-community==0.0.6
langchain-core==0.0.11
langchain-chroma==0.0.4
chromadb==0.4.13
pydantic==1.10.11
sqlalchemy==2.0.20
requests==2.31.0
urllib3==2.0.4
numpy==1.25.2
pandas==2.0.3
sentence-transformers==2.2.2
transformers==4.32.1
torch==2.0.1
#export HUGGINGFACE_API_TOKEN="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"